lab_1/requirements.txt
numpy==2.2.4
pygame==2.6.1
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
